+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.3] - Unreleased
 ### Added
 - TimeOrderMixin for models that have time_order columns
+- Integration tests refer to `MYSQL_CONNECTION_URI` environment variable to connect to the DB
+- This changelog
 
 ### Changed
 - Moved common fixtures to conftest.py